remove add accounts page
